+++ jwst/source_catalog/source_catalog_step.py
@@ -6,11 +6,11 @@ import os
 import warnings
 
 from crds.core.exceptions import CrdsLookupError
+import numpy as np
 from photutils.utils.exceptions import NoDetectionsWarning
 
 from .source_catalog import (ReferenceData, Background, make_kernel,
-                             make_segment_img, calc_total_error,
-                             JWSTSourceCatalog)
+                             make_segment_img, JWSTSourceCatalog)
 from .. import datamodels
 from ..stpipe import Step
 
@@ -72,10 +72,10 @@ class SourceCatalogStep(Step):
                 self.log.warning(msg)
                 return
 
-            coverage_mask = (model.wht == 0)
+            coverage_mask = np.isnan(model.err)
             if coverage_mask.all():
-                self.log.warning('There are no pixels with non-zero weight. '
-                                 'Source catalog will not be created.')
+                self.log.warning('There are no valid pixels. Source catalog '
+                                 'will not be created.')
                 return
 
             bkg = Background(model.data, box_size=self.bkg_boxsize,
@@ -99,13 +99,9 @@ class SourceCatalogStep(Step):
                 return
             self.log.info(f'Detected {segment_img.nlabels} sources')
 
-            # TODO: update when model contains errors
-            total_error = calc_total_error(model)
-
             ci_star_thresholds = (self.ci1_star_threshold,
                                   self.ci2_star_threshold)
-            catobj = JWSTSourceCatalog(model, segment_img, error=total_error,
-                                       kernel=kernel,
+            catobj = JWSTSourceCatalog(model, segment_img, kernel=kernel,
                                        kernel_fwhm=self.kernel_fwhm,
                                        aperture_params=aperture_params,
                                        abvega_offset=abvega_offset,